Move code from DeviceService to ExportManager
