+++ src/AzureIoTHub.Portal.Server/Services/DeviceServiceBase.cs
@@ -19,14 +19,13 @@ namespace AzureIoTHub.Portal.Server.Services
     using Mappers;
     using Shared.Models.v1._0.Filters;
     using Device = Domain.Entities.Device;
-    using System.IO;
 
     public abstract class DeviceServiceBase<TDto> : IDeviceService<TDto>
         where TDto : IDeviceDetails
     {
-        protected readonly PortalDbContext portalDbContext;
+        private readonly PortalDbContext portalDbContext;
         private readonly IExternalDeviceService externalDevicesService;
-        protected readonly IDeviceTagService deviceTagService;
+        private readonly IDeviceTagService deviceTagService;
         private readonly IDeviceModelImageManager deviceModelImageManager;
         private readonly IDeviceTwinMapper<DeviceListItem, TDto> deviceTwinMapper;
 
@@ -207,7 +206,5 @@ namespace AzureIoTHub.Portal.Server.Services
                     .Select(tag => new { tag, tagValue = tags[tag] })
                     .ToDictionary(pair => pair.tag, pair => pair.tagValue);
         }
-
-        public abstract Task<Stream> ExportDeviceList();
     }
 }
